sync revision 2219
Build:
	- support for building with CMake on all supported desktop platforms
	- headers can now be compiled by a C++ compiler

Platform/X11:
	- use XInput2 if available

Platform/Win32:
	- Windows 11 immersive dark mode

Engine:
	- plugin support
	- replaced math library with DirectXMath
	- Lua functions can now be registered as event handlers
	- the startup scene can now be specified on the command line (-s <path>)

Editor:
	- Assimp importer
	- replaced platform-specific GUI implementations with Qt6/Qt5
	- basic inspector functionality
	- play/pause (starts a new engine instance)

Plugins:
	- TTS plugin

Deps:
	- cgltf 1.13
	- meshoptimizer 0.18
	- volk 1.3.226

+++ Deps/jsmn.h
@@ -221,7 +221,7 @@ static int jsmn_parse_string(jsmn_parser *parser, const char *js,
     }
 
     /* Backslash: Quoted symbol expected */
-    if (c == '\\' && parser->pos + 1 < len) {
+    if (c == '\\' && (size_t)parser->pos + 1 < len) {
       int i;
       parser->pos++;
       switch (js[parser->pos]) {
